exercise: determine the mcl particle weights
Particle sensor measurements are taken for each of the simulated particles. A probability is returned from a gaussian pdf which is good for relative probability only. These probabilities are generated assuming that the distance to a landmark from a particle is the true mean, the sensor noise attributed to the particle is the standard deviation, and the actual robot measurement (for the given landmark) is the observable value, typically called x.