exercise: create a function which will return probability
The input vector is a list of numbers. The main script calculates the sum for the vector. This is used with each element of the vector to calculate a probability for the element. Note that this is in the mcl context where alphas need to be created for the weights vector that is calculated from the normal pdf.