exercise: plotting of the map generated by the occupancy grid algorithm implementation
The program takes inputs of sensor measurements, and robot poses. The occupancy grid algorithm is applied which checks all grid cells. If a grid cell is in the sensing radius, then the binary bayes filter is applied which uses the inverse sensor model if the cell is within the sensor radius. The plotting is done using matplotlibcpp.